backend/requirements.txt
python-dotenv==1.1.1
django 
djangorestframework 
psycopg2-binary 
django-cors-headers
djangorestframework-simplejwt
qrcode[pil]
Pillow
